Remove tracked files ignored by .gitignore

+++ FlowerEcommerce.Domain/Entities/OrderItem.cs
@@ -7,10 +7,10 @@ namespace FlowerEcommerce.Domain.Entities
 {
     public class OrderItem : ModificationAuditedEntity
     {
-        public int ProductId { get; set; }
+        public ulong ProductId { get; set; }
         public Product Product { get; set; }
 
-        public int OrderId { get; set; }
+        public ulong OrderId { get; set; }
         public Order Order { get; set; }
 
         public int Quantity { get; set; }
